Fixed PyQt4 configure.py arguments
I: Extracting PyQt4_gpl_x11-4.12.tar.gz to PyQt4_gpl_x11-4.12
Determining the layout of your Qt installation...
This is the GPL version of PyQt 4.12 (licensed under the GNU General Public
License) for Python 2.7.10 on darwin.
Found the license file pyqt-gpl.sip.
Checking to see if the QtGui module should be built...
Checking to see if the QtHelp module should be built...
Checking to see if the QtMultimedia module should be built...
Checking to see if the QtNetwork module should be built...
Checking to see if the QtDBus module should be built...
Checking to see if the QtDeclarative module should be built...
Checking to see if the QtScript module should be built...
Checking to see if the QtScriptTools module should be built...
Checking to see if the QtXml module should be built...
Checking to see if the QtOpenGL module should be built...
Checking to see if the QtSql module should be built...
Checking to see if the QtSvg module should be built...
Checking to see if the QtTest module should be built...
Checking to see if the QtWebKit module should be built...
Checking to see if the QtXmlPatterns module should be built...
Checking to see if the phonon module should be built...
Checking to see if the QtAssistant module should be built...
Checking to see if the QtDesigner module should be built...
Checking to see if the QAxContainer module should be built...
Checking to see if the dbus support module should be built...
DBus v1 does not seem to be installed.
Qt v4.8.7 free edition is being used.
SIP 4.19.1 is being used.
The Qt header files are in
/Applications/GNURadio.app/Contents/MacOS/usr/include.
The shared Qt libraries are in
/Applications/GNURadio.app/Contents/MacOS/usr/lib.
The Qt binaries are in /Applications/GNURadio.app/Contents/MacOS/usr/bin.
The Qt mkspecs directory is in /Applications/GNURadio.app/Contents/MacOS/usr.
These PyQt modules will be built: QtCore, QtGui, QtHelp, QtNetwork, QtXml,
QtSql, QtSvg, QtTest, QtDesigner.
The PyQt Python package will be installed in
/Applications/GNURadio.app/Contents/MacOS/usr/lib/python2.7/site-packages.
PyQt is being built with generated docstrings.
PyQt is being built with 'protected' redefined as 'public'.
The Designer plugin will be installed in
/Applications/GNURadio.app/Contents/MacOS/usr/plugins/designer.
The PyQt .sip files will be installed in
/Applications/GNURadio.app/Contents/MacOS/usr/share/sip.
pyuic4, pyrcc4 and pylupdate4 will be installed in
/Applications/GNURadio.app/Contents/MacOS/usr/bin.
Embedding sip flags...
Generating the C++ source for the QtCore module...
Creating the Makefile for the QtCore module...
Generating the C++ source for the QtGui module...
Creating the Makefile for the QtGui module...
Generating the C++ source for the QtHelp module...
Creating the Makefile for the QtHelp module...
Generating the C++ source for the QtNetwork module...
Creating the Makefile for the QtNetwork module...
Generating the C++ source for the QtSql module...
Creating the Makefile for the QtSql module...
Generating the C++ source for the QtSvg module...
Creating the Makefile for the QtSvg module...
Generating the C++ source for the QtTest module...
Creating the Makefile for the QtTest module...
Generating the C++ source for the QtXml module...
Creating the Makefile for the QtXml module...
Generating the C++ source for the QtDesigner module...
Creating the Makefile for the QtDesigner module...
Generating the C++ source for the Qt module...
Creating the Makefile for the Qt module...
Creating QScintilla API file...
Creating top level Makefile...
Creating QPy support library for QtCore Makefile...
Creating QPy support library for QtGui Makefile...
Creating QPy support library for QtDesigner Makefile...
Creating QPy support libraries Makefile...
Creating pyuic4 wrapper...
Creating pyuic4 Makefile...
Creating pylupdate4 Makefile...
Creating pyrcc4 Makefile...
Creating Qt Designer plugin Makefile...
Creating pyqtconfig.py...